Moved progress bar to printers. Pretty printer prints test asynchronously.
