+++ printers/progress_bar.go
@@ -1,4 +1,4 @@
-package main
+package printers
 
 import (
 	"fmt"
